+++ recipes/ring-span-lite/all/conanfile.py
@@ -8,12 +8,13 @@ required_conan_version = ">=1.50.0"
 
 class RingSpanLiteConan(ConanFile):
     name = "ring-span-lite"
-    url = "https://github.com/conan-io/conan-center-index"
-    homepage = "https://github.com/martinmoene/ring-span-lite"
     description = (
         "ring-span lite - A ring_span type for C++98, C++11 and later in a single-file header-only library ")
-    topics = ("cpp98", "cpp11", "cpp14", "cpp17", "ring-span")
     license = "BSL-1.0"
+    url = "https://github.com/conan-io/conan-center-index"
+    homepage = "https://github.com/martinmoene/ring-span-lite"
+    topics = ("cpp98", "cpp11", "cpp14", "cpp17", "ring-span", "header-only")
+    package_type = "header-library"
     settings = "os", "arch", "compiler", "build_type"
     no_copy_source = True
 
@@ -24,11 +25,7 @@ class RingSpanLiteConan(ConanFile):
         basic_layout(self, src_folder="src")
 
     def source(self):
-        get(self, **self.conan_data["sources"][self.version],
-            destination=self.source_folder, strip_root=True)
-
-    def build(self):
-        pass
+        get(self, **self.conan_data["sources"][self.version], strip_root=True)
 
     def package(self):
         copy(self, "*.hpp", src=os.path.join(self.source_folder, "include"),