+++ docs/changelog.md
@@ -1,5 +1,25 @@
 # Changelog
 
+### 12-Sep-2024 - 09:23 CEST
+
+- [feature] Add support for Conan 2.7.1 in the CI
+- [feature] Add support for Conan 1.65.0 in the CI
+
+### 10-Jul-2024 - 13:22 CEST
+
+- [feature] Add support for Conan 2.5.0 in the CI
+- [fix] Invalid configuration from tool requirement in Conan 1.x
+
+### 22-May-2024 - 12:04 CEST
+
+- [feature] Add support for Conan 2.3.1 in the CI
+- [feature] Add support for Conan 1.64.1 in the CI
+
+### 03-April-2024 - 13:53 CEST
+
+- [feature] Add support for Conan 2.2.2 in the CI
+- [feature] Add support for Conan 1.63.0 in the CI
+
 ### 20-March-2024 - 11:13 CET
 
 - [fix] Changing Version Ranges in dependencies is now bump dependencies